Looks like a big refactor but basically moved all code to a new file
